refactor: isAddNewAddressVisible ref and temp dir for components
